Load `theme-setup-dialog` from the correct location.
